requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.12
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.3.0
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.33
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
pydantic==2.13.3
pydantic_core==2.46.3
pyjokes==0.8.3
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
langchain-groq==1.1.2
python-dotenv==1.2.2